sql: move decorrelation mod into its own file
Having large inline modules is confusing, in my opinion, and introduces
an extra layer of nesting that makes rustfmt format things poorly.
